copying folders lead to crash

+++ lsl_autobids/convert_to_bids_and_upload.py
@@ -124,7 +124,8 @@ class BIDS:
                 for file in os.listdir(experiments_path):
                     src_file = os.path.join(experiments_path, file)
                     dest_file = os.path.join(dest_dir, file)
-                    shutil.copy(src_file, dest_file)
+                    if os.path.isfile(src_file):
+                        shutil.copy(src_file, dest_file)
 
                 # Compress the 'other' directory into a ZIP file
                 shutil.make_archive(dest_dir, 'zip', dest_dir)
